Add `guid` and `release_license` columns on `release` table.

+++ src/main/java/org/oagi/srt/gateway/http/api/release_management/data/SimpleRelease.java
@@ -1,12 +1,13 @@
 package org.oagi.srt.gateway.http.api.release_management.data;
 
 import lombok.Data;
-import org.oagi.srt.entity.jooq.enums.ReleaseState;
+
+import java.math.BigInteger;
 
 @Data
 public class SimpleRelease {
 
-    private long releaseId;
+    private BigInteger releaseId;
     private String releaseNum;
     private ReleaseState state;
